Replace a split method by a StringTokenizer clas to parse run param.
In current implementation for split "Run Paramteres" of project
is used split method that split a String by space. I that not handle
qouted parameters with spaces and split it by space that not allow
to pass the text as parameter with spaces.

To handle this situation the split method are replaced by use the
StringTokenizer class, that support single (`) and duble (") quote
that arrond the tekst as a params.

Task: #123